// api/consolidated/models/DocumentOperationResult.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { Document } from './Document';
import type { EntityValidationError } from './EntityValidationError';

export type DocumentOperationResult = {
    outcome?: DocumentOperationResult.outcome;
    errors?: Array<EntityValidationError> | null;
    data?: Document;
    correlationId?: string | null;
    status?: number;
    message?: string | null;
    trace?: string | null;
};

export namespace DocumentOperationResult {

    export enum outcome {
        SUCCESS = 'Success',
        FAILURE = 'Failure',
    }


}